Provide a reload button on saved searches
Included is a fix for a bug in titlebox code when multiple
buttons are providing that caused them to be stacked on top
of each other. The "right" class needed to be on a span around
all of the provided buttons.

Also show the load time of searches even when refresh isn't
enabled because it's always useful to see how old the information
is. With the refresh button and reveal options, the load time can
be different from when the main page was loaded.