Don't Show Avatar when ShowUser is called from Results.tsv
Tests fails because the ShowUser element is used in the Results.tsv
template, which is used to generate a TSV (or XLSX) file.

The ShowUser element is used to display the user's avatar, which is
not supported in the TSV (or XLSX) format.

This commit adds a check to not display the avatar when the ShowUser
element is called from the Results.tsv template.