Allow users choose if they want to see Avatars or not
Now that we have the ability to show user initials and avatars, we should
allow users to choose if they want to see avatars or not sytem-wide by
adding a new options to the UsernameFormat configuration option.

It adds a new possible return to RT::User->Format, which is a boolean
value that indicates if the user wants to see avatars or not. This is
only returned if the context that is calling the method expects an array
as a return.

For that reason we also update some parts of RT code and tests be
compatible with this new return.